+++ Changelog.rst
@@ -1,3 +1,12 @@
+version NEXT
+------------
+----
+
+**2021-??-??**
+
+* Fix for zero-sized unlimited dimensions when read from a grouped
+  netCDF file (https://github.com/NCAS-CMS/cfdm/issues/113)
+
 version 1.8.8.0
 ---------------
 ----